+++ extensions/gdx-ai/.classpath
