+++ extensions/gdx-ai/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>gdx-aida</name>
+	<name>gdx-ai</name>
 	<comment></comment>
 	<projects>
 	</projects>
