+++ extensions/gdx-ai/.settings/org.eclipse.jdt.core.prefs
