+++ extensions/gdx-ai/src/com/badlogic/gdx/ai/Agent.java
@@ -14,9 +14,9 @@
  * limitations under the License.
  ******************************************************************************/
 
-package com.badlogic.gdx.aida;
+package com.badlogic.gdx.ai;
 
-import com.badlogic.gdx.aida.msg.Telegram;
+import com.badlogic.gdx.ai.msg.Telegram;
 
 /** @author davebaol */
 public interface Agent {
