+++ extensions/gdx-ai/src/com/badlogic/gdx/ai/fsm/DefaultStateMachine.java
@@ -14,9 +14,9 @@
  * limitations under the License.
  ******************************************************************************/
 
-package com.badlogic.gdx.aida.fsm;
+package com.badlogic.gdx.ai.fsm;
 
-import com.badlogic.gdx.aida.msg.Telegram;
+import com.badlogic.gdx.ai.msg.Telegram;
 
 /** Default implementation of the {@link StateMachine} interface.
  * @author davebaol */
