+++ extensions/gdx-ai/src/com/badlogic/gdx/ai/fsm/State.java
@@ -14,9 +14,9 @@
  * limitations under the License.
  ******************************************************************************/
 
-package com.badlogic.gdx.aida.fsm;
+package com.badlogic.gdx.ai.fsm;
 
-import com.badlogic.gdx.aida.msg.Telegram;
+import com.badlogic.gdx.ai.msg.Telegram;
 
 /** The state of a state machine defines the logic of the entities that enter, exit and last this state. Additionally, a state may
  * be delegated by an entity to handle its messages.
