+++ tests/gdx-tests/.classpath
@@ -7,6 +7,6 @@
 	<classpathentry combineaccessrules="false" exported="true" kind="src" path="/gdx-bullet"/>
 	<classpathentry combineaccessrules="false" exported="true" kind="src" path="/gdx-freetype"/>
 	<classpathentry combineaccessrules="false" kind="src" path="/gdx-controllers"/>
-	<classpathentry combineaccessrules="false" kind="src" path="/gdx-aida"/>
+	<classpathentry combineaccessrules="false" kind="src" path="/gdx-ai"/>
 	<classpathentry kind="output" path="bin"/>
 </classpath>
